chore: move skeleton animation to animations utils/folder
