+++ src/app/utils/animations/useSkeletonAnimation.tsx
